moved test, and moved caching
