+++ packages/drivers/odsp-driver/src/test/createNewUtilsTests.spec.ts
@@ -5,52 +5,60 @@
 
 import { strict as assert } from "assert";
 import * as api from "@fluidframework/protocol-definitions";
+import { TelemetryNullLogger } from "@fluidframework/common-utils";
+import { IOdspResolvedUrl } from "@fluidframework/odsp-driver-definitions";
+import { getDocAttributesFromProtocolSummary } from "@fluidframework/driver-utils";
 import { convertSummaryTreeToIOdspSnapshot } from "../createNewUtils";
+import { createNewFluidFile } from "../createFile";
 import { IOdspSnapshotTreeEntryTree } from "../contracts";
+import { EpochTracker } from "../epochTracker";
+import { getHashedDocumentId } from "../odspPublicUtils";
+import { INewFileInfo, createCacheSnapshotKey } from "../odspUtils";
+import { LocalPersistentCache } from "../odspCache";
+import { mockFetchOk } from "./mockFetch";
+
+const createUtLocalCache = () => new LocalPersistentCache(2000);
 
 describe("Create New Utils Tests", () => {
+    const createSummary = () => {
+        const summary: api.ISummaryTree = {
+            type: api.SummaryType.Tree,
+            tree: {},
+        };
+
+        summary.tree[".app"] = {
+            type: api.SummaryType.Tree,
+            tree: {
+                attributes: {
+                    type: api.SummaryType.Blob,
+                    content: "testing",
+                },
+            },
+        };
+        summary.tree[".protocol"] = {
+            type: api.SummaryType.Tree,
+            tree: {
+                attributes: {
+                    type: api.SummaryType.Blob,
+                    content: JSON.stringify({ branch: "", minimumSequenceNumber: 0, sequenceNumber: 0,
+                        term: 1 }),
+                },
+            },
+        };
+        return summary;
+    };
+
     beforeEach(() => {
     });
 
     it("Should convert as expected and check contents", async () => {
-        const rootBlob: api.ISummaryBlob = {
-            type: api.SummaryType.Blob,
-            content: JSON.stringify("root"),
-        };
-        const componentBlob: api.ISummaryBlob = {
-            type: api.SummaryType.Blob,
-            content: JSON.stringify("component"),
-        };
-        const contentBlob: api.ISummaryBlob = {
-            type: api.SummaryType.Blob,
-            content: "[]",
-        };
         const rootBlobPath = "default/root";
         const componentBlobPath = "default/component";
         const contentBlobPath = "contentTree/contentBlob";
-        const appSummary: api.ISummaryTree = {
-            type: api.SummaryType.Tree,
-            tree: {
-                default: {
-                    type: api.SummaryType.Tree,
-                    tree: {
-                        component: componentBlob,
-                        root: rootBlob,
-                    },
-                },
-                contentTree: {
-                    type: api.SummaryType.Tree,
-                    tree: {
-                        contentBlob,
-                    },
-                    unreferenced: true,
-                },
-            },
-        };
 
-        const odspSnapshot = convertSummaryTreeToIOdspSnapshot(appSummary);
+        const odspSnapshot = convertSummaryTreeToIOdspSnapshot(createSummary());
         assert.strictEqual(odspSnapshot.trees.length, 1, "1 main tree should be there");
-        assert.strictEqual(odspSnapshot.blobs?.length, 3, "3 blobs should be there");
+        assert.strictEqual(odspSnapshot.blobs?.length, 2, "2 blobs should be there");
 
         const mainTree = odspSnapshot.trees[0];
         assert.strictEqual(mainTree.id, odspSnapshot.id, "Main tree id should match");
@@ -66,7 +74,7 @@ describe("Create New Utils Tests", () => {
         });
 
         // Validate that the snapshot has all the expected blob entries.
-        assert.strictEqual(blobEntries.length, 3, "There should be 3 blob entries in the main tree");
+        assert.strictEqual(blobEntries.length, 2, "There should be 2 blob entries in the main tree");
         assert(blobEntries.includes(rootBlobPath), "Root blob should exist");
         assert(blobEntries.includes(componentBlobPath), "Component blob should exist");
         assert(blobEntries.includes(contentBlobPath), "Content blob should exist");
@@ -80,5 +88,51 @@ describe("Create New Utils Tests", () => {
                 assert(treeEntry.unreferenced, "content tree entry should be unreferenced");
             }
         }
+    });
+    it("Should cache converted summary during createNewFluidFile", async () => {
+        const siteUrl = "https://microsoft.sharepoint-df.com/siteUrl";
+        const driveId = "driveId";
+        const itemId = "itemId";
+        const hashedDocumentId = getHashedDocumentId(driveId, itemId);
+        const resolvedUrl = ({ siteUrl, driveId, itemId, odspResolvedUrl: true } as any) as IOdspResolvedUrl;
+        const localCache = createUtLocalCache();
+        // use null logger here as we expect errors
+        const epochTracker = new EpochTracker(
+            localCache,
+            {
+                docId: hashedDocumentId,
+                resolvedUrl,
+            },
+            new TelemetryNullLogger());
+
+        const filePath = "path";
+        const newFileParams: INewFileInfo = {
+            driveId,
+            siteUrl: "https://www.localhost.xxx",
+            filePath,
+            filename: "filename",
+        };
+
+        const odspResolvedUrl = await mockFetchOk(
+                async () =>createNewFluidFile(
+                    async (_options) => "token",
+                    newFileParams,
+                    new TelemetryNullLogger(),
+                    createSummary(),
+                    epochTracker,
+                ) ,
+                { itemId: "itemId1"},
+                { "x-fluid-epoch": "epoch1" },
+                );
+        const value = await epochTracker.get(createCacheSnapshotKey(odspResolvedUrl));
+        const blobs = value.snapshot.blobs;
+        assert.strictEqual(blobs.length, 2, "wrong length of blobs");
+        assert.strictEqual(blobs[0].content, "testing", "wrong content of testing blob");
+
+        const protocolSummary = createSummary().tree[".protocol"] as api.ISummaryTree;
+        const documentAttributes = getDocAttributesFromProtocolSummary(protocolSummary);
+        assert.strictEqual(documentAttributes.minimumSequenceNumber, 0, "wrong min sequence number");
+        assert.strictEqual(documentAttributes.sequenceNumber, 0, "wrong sequence number");
+        assert.strictEqual(documentAttributes.term, 1, "wrong term");
     });
 });
